Move common function to 'src/packages'
